fix: capitalize constants names (B7), move types to a separate file
