+++ src/components/app/app.tsx
@@ -6,7 +6,8 @@ import Login from '../../pages/login/login';
 import Offer from '../../pages/offer/offer';
 import NotFound from '../../pages/not-found/not-found';
 import PrivateRoute from '../private-route/private-route';
-import {Offer as Place, Comment, AppRoute } from '../../const';
+import { Offer as Place, Comment } from '../../types';
+import { APP_ROUTE } from '../../const';
 
 type AppProps = {
   offersCount: number;
@@ -14,27 +15,34 @@ type AppProps = {
   comments: Comment[];
 };
 
-export default function App({ places, comments, offersCount }: AppProps): JSX.Element {
+export default function App({
+  places,
+  comments,
+  offersCount,
+}: AppProps): JSX.Element {
   return (
     <BrowserRouter>
       <Routes>
-        <Route path={AppRoute.Root} element={<Layout />}>
+        <Route path={APP_ROUTE.ROOT} element={<Layout />}>
           <Route
             index
             element={<Main offersCount={offersCount} places={places} />}
           />
           <Route
-            path={AppRoute.Favorites}
+            path={APP_ROUTE.FAVORITES}
             element={
               <PrivateRoute>
                 <Favorites />
               </PrivateRoute>
             }
           />
-          <Route path={AppRoute.Offer} element={<Offer comments={comments} />} />
+          <Route
+            path={APP_ROUTE.OFFER}
+            element={<Offer comments={comments} />}
+          />
         </Route>
-        <Route path={AppRoute.Login} element={<Login />} />
-        <Route path={AppRoute.NotFound} element={<NotFound />} />
+        <Route path={APP_ROUTE.LOGIN} element={<Login />} />
+        <Route path={APP_ROUTE.NOT_FOUND} element={<NotFound />} />
       </Routes>
     </BrowserRouter>
   );
