+++ src/pages/offer/offer.tsx
@@ -1,6 +1,6 @@
 import ReviewForm from '../../components/review-form/review-form';
 import ReviewsList from '../../components/reviews-list/reviews-list';
-import { Comment } from '../../const';
+import { Comment } from '../../types';
 
 type OfferProps = {
   comments: Comment[];
